[REF] Extract unfold functionality to separate file
Unfold functionality is stored in ``unfoldNd/unfold.py``.
Fold functionality will be implemented in ``unfoldNd/fold.py``
