+++ test/utils_settings.py
@@ -1,0 +1,7 @@
+"""Problem settings for utils test."""
+
+UNSUPPORTED_N = [4, -1]
+UNSUPPORTED_N_IDS = [f"N={n}" for n in UNSUPPORTED_N]
+
+UNSUPPORTED_KERNEL_SIZE = [[1, 2], 1]
+UNSUPPORTED_KERNEL_SIZE_IDS = [f"kernel_size={s}" for s in UNSUPPORTED_KERNEL_SIZE]
